+++ custom_components/energy_management_system/config_flow.py
@@ -46,7 +46,9 @@ from .const import (
     CONF_WATER_SOFTENER_START_HOUR,
     CONF_PV_ENERGY_SENSOR,
     CONF_PV_ACTUAL_TILT_DEGREES,
+    CONF_DISHWASHER_START_IN,
     CONF_SUN_AZIMUTH_SENSOR,
+    CONF_WASHING_MACHINE_END_AT,
     CONF_SUN_ELEVATION_SENSOR,
     CONF_SUN_PHASE_SENSOR,
     CONF_BATTERY_MODULE_CELL_VOLTAGE_MAX_SENSORS,
@@ -416,6 +418,18 @@ def _schema(defaults: dict | None = None) -> vol.Schema:
                 CONF_SUN_AZIMUTH_SENSOR,
                 default=defaults.get(CONF_SUN_AZIMUTH_SENSOR),
             ): selector.EntitySelector(selector.EntitySelectorConfig(domain="sensor")),
+            vol.Optional(
+                CONF_DISHWASHER_START_IN,
+                default=defaults.get(CONF_DISHWASHER_START_IN),
+            ): selector.EntitySelector(
+                selector.EntitySelectorConfig(domain=["number", "sensor"])
+            ),
+            vol.Optional(
+                CONF_WASHING_MACHINE_END_AT,
+                default=defaults.get(CONF_WASHING_MACHINE_END_AT),
+            ): selector.EntitySelector(
+                selector.EntitySelectorConfig(domain=["sensor", "number"])
+            ),
             vol.Optional(
                 CONF_SUN_PHASE_SENSOR,
                 default=defaults.get(CONF_SUN_PHASE_SENSOR),